All Validation work finished

+++ LogeenStockManagement/Controllers/SaleBillsController.cs
@@ -45,7 +45,7 @@ namespace LogeenStockManagement.Controllers
         [HttpPut("{id}")]
         public async Task<IActionResult> PutSaleBill(int id, SaleBill saleBill)
         {
-            if (id != saleBill.Id)
+            if (id != saleBill.Id|| IsSaleBillDataNotValid(saleBill))
             {
                 return BadRequest();
             }
@@ -72,16 +72,21 @@ namespace LogeenStockManagement.Controllers
         }
 
         // POST: api/SaleBills
-        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPost]
         public async Task<ActionResult<SaleBill>> PostSaleBill(SaleBill saleBill)
         {
+            if (IsSaleBillDataNotValid(saleBill))
+            {
+                return BadRequest();
+            }
+
             _context.SaleBills.Add(saleBill);
             await _context.SaveChangesAsync();
 
             return CreatedAtAction("GetSaleBill", new { id = saleBill.Id }, saleBill);
         }
 
+        //TODO remove delete functionality
         // DELETE: api/SaleBills/5
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteSaleBill(int id)
@@ -90,6 +95,14 @@ namespace LogeenStockManagement.Controllers
             if (saleBill == null)
             {
                 return NotFound();
+            }
+
+            if (saleBill.ImportPayments.Count > 0 ||
+                saleBill.SaleBillProducts.Count > 0 ||
+                saleBill.SalesReturnsBills.Count > 0
+               )
+            {
+                return BadRequest();
             }
 
             _context.SaleBills.Remove(saleBill);
@@ -102,5 +115,58 @@ namespace LogeenStockManagement.Controllers
         {
             return _context.SaleBills.Any(e => e.Id == id);
         }
+
+        public bool IsSaleBillDataNotValid(SaleBill SaleBill)
+        {
+            /*
+             * ID INT IDENTITY(1,1),
+              BillCode VARCHAR(50) UNIQUE,
+              [Date] DATE NOT NULL Default GETDATE(),
+              BillType VARCHAR(10) NOT NULL, --TODO check in cash or Debit
+              Discount FLOAT NOT NULL,
+              CheckNumber VARCHAR(100) NOT NULL,
+              paidup FLOAT,
+              Remaining FLOAT ,
+              BillTotalPrice FLOAT,
+              StockId INT NOT NULL,
+              PayMethodId INT NOT NULL,
+              ClientId INT,
+              TaxId INT,
+              CONSTRAINT SaleBillPK PRIMARY KEY (ID),
+              CONSTRAINT SaleStockFK FOREIGN KEY (StockId) REFERENCES Stock(ID),
+              CONSTRAINT SalePayMethodFK FOREIGN KEY (PayMethodId) REFERENCES PaymentMethod(ID),
+              CONSTRAINT SaleClientFK FOREIGN KEY (ClientId) REFERENCES Client(ID),
+              CONSTRAINT SaleTaxFK FOREIGN KEY (TaxId) REFERENCES Tax(ID),
+              Constraint BillTypeCheck check(BillType in ('Cash','Debit') )
+             */
+            //foreign keys can not refer to Not Existed
+            bool StockExisted = _context.Stocks.Any(s => s.Id == SaleBill.StockId);
+            bool taxExisted = _context.Taxes.Any(t => t.Id == SaleBill.TaxId);
+            bool PayMethodExisted = _context.PaymentMethods.Any(p => p.Id == SaleBill.PayMethodId);
+            bool ClientExisted = _context.Clients.Any(c => c.Id == SaleBill.ClientId);
+
+            // UNIQUE Prorerty must to be Not Existed before
+            bool BillCodeRepeat = _context.SaleBills.Any((b) => b.BillCode == SaleBill.BillCode && b.Id != SaleBill.Id);
+            bool BillTypevalid = SaleBill.BillType == "Cash" || SaleBill.BillType == "Debit";
+
+            //Not NUll properties + chech Foreign and Uniqe result
+            if (
+                SaleBill.BillCode == null || SaleBill.CheckNumber == null || SaleBill.BillTotalPrice == 0 ||
+                !StockExisted || !taxExisted || !PayMethodExisted || !ClientExisted ||
+                !BillTypevalid || BillCodeRepeat||
+                !DateTime.TryParse(SaleBill.Date.ToString(),out _)
+                )
+            {
+                return true;
+            }
+            else
+            {
+                return false;
+            }
+
+        }
+
+
+
     }
 }
